+++ simplexML/Pyscripts/rse.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""
+Auxiliar function to compute Root Squared Error
+"""
+import numpy as np
+
+def calc_rse(valores,prediccion):
+    return(sum((valores-prediccion)**2)/sum((valores-np.mean(valores))**2))
