+++ simplexML/tables/Model_Errors.csv
@@ -1,0 +1,19 @@
+"nexper";"model";"Method";"Index";"Min";"Q1";"Median";"Mean";"Q3";"Max"
+100;"ABIOTIC";"Linear Regressor";"RSE";0,970788298817531;0,974153447462179;0,975305476769996;0,975747967685454;0,976960120097097;0,983368435969895
+100;"ABIOTIC";"Linear Regressor";"RMSE";34,9131436096216;37,8065868296414;38,9327814654475;38,9321860960878;39,8608917833418;42,9490352406201
+100;"TWOSTEP";"Linear Regressor";"RSE";0,705215142155936;0,741385965090606;0,75553483198233;0,757470185793235;0,771500822882754;0,83204412807145
+100;"TWOSTEP";"Linear Regressor";"RMSE";42,3576000489103;46,2959818725214;48,0445939335361;48,1394923332977;49,8926158102606;56,0087793693155
+100;"ALLFEATURES";"Linear Regressor";"RSE";0,875489838725253;0,891763729466147;0,897897480993024;0,898593932143415;0,904319953204162;0,931929506227071
+100;"ALLFEATURES";"Linear Regressor";"RMSE";34,8831240811559;36,8025118575265;37,4582710322306;37,5635756384295;38,3418753268409;41,1238550748886
+100;"ABIOTIC";"Random Forest";"RSE";0,111285185215552;0,132781196743351;0,138499133737026;0,140437535530875;0,146752200214439;0,189646811114981
+100;"ABIOTIC";"Random Forest";"RMSE";13,1837573409517;14,3720895213916;14,7027458294966;14,7440173880507;15,1546846978545;16,7826740110536
+100;"TWOSTEP";"Random Forest";"RSE";0,0387250429444789;0,061603811122769;0,0693529709289709;0,0714737313045009;0,0808512338897158;0,112535900262317
+100;"TWOSTEP";"Random Forest";"RMSE";10,9256401610505;13,7930994247927;14,5712039107145;14,6746503684001;15,5431030272737;18,3128040438274
+100;"ALLFEATURES";"Random Forest";"RSE";0,0992603047177586;0,123504403448371;0,129942305315553;0,130982742269662;0,138228123095583;0,156841491075857
+100;"ALLFEATURES";"Random Forest";"RMSE";12,6950445357351;13,8518405045558;14,2943726365968;14,3187240632064;14,864963452696;16,1542317109459
+100;"ABIOTIC";"XGBoost";"RSE";0,137744408614216;0,158606378381508;0,166152317687748;0,168253364140975;0,174687296062962;0,211312846545217
+100;"ABIOTIC";"XGBoost";"RMSE";14,156226182313;15,7152938793976;16,1162118665397;16,1414479670316;16,531006221651;17,9095541129973
+100;"TWOSTEP";"XGBoost";"RSE";0,0423142557463917;0,0685910315363641;0,0774717184243325;0,0790161116815812;0,0904856111414382;0,113885304360818
+100;"TWOSTEP";"XGBoost";"RMSE";12,2677510997765;14,5635843143551;15,7617141630238;15,4423754195562;16,3088951073931;18,9150189858701
+100;"ALLFEATURES";"XGBoost";"RSE";0,105824150874145;0,126014156893932;0,133645790003193;0,133941041756289;0,142680776188089;0,156399093939172
+100;"ALLFEATURES";"XGBoost";"RMSE";13,0481504214863;14,0808413553381;14,391525213128;14,4828847293943;14,9756614968149;16,1287924215033
