+++ simplexML/tables/datos_err_onlyspecies.csv
@@ -1,0 +1,70 @@
+"Species";"Method";"RMSE";"RSE"
+"SPRU";"Linear_Regression";20,6431714349718;54,4590475088134
+"MEPO";"Linear_Regression";23,1746788827915;3,76689559985906
+"FRPU";"Linear_Regression";24,0991034090226;3,40169761554111
+"MEEL";"Linear_Regression";24,1566421842117;3,67410988200917
+"POMA";"Linear_Regression";26,8369965794068;1,45097822324657
+"HOMA";"Linear_Regression";145,288070447873;1,09434110861247
+"SUSP";"Linear_Regression";21,385094435827;893,082053535317
+"RAPE";"Linear_Regression";24,3778479483812;4,87117019787144
+"POMO";"Linear_Regression";23,9270887232681;6,68742880932236
+"SCLA";"Linear_Regression";23,0784797345457;31,5848838748206
+"SASO";"Linear_Regression";31,3717928737671;1,4185669269983
+"CHFU";"Linear_Regression";31,9547103657549;2,54501659513412
+"PLCO";"Linear_Regression";27,562230611644;1,68492850883715
+"BEMA";"Linear_Regression";30,0663533383632;35,9539843724639
+"CETE";"Linear_Regression";28,8679162134225;16,8392557472625
+"LEMA";"Linear_Regression";51,5205340644781;0,95646799626255
+"MESU";"Linear_Regression";25,6477512016461;4,57257166468131
+"LYTR";"Linear_Regression";24,7750569865511;2,45817991492484
+"COSQ";"Linear_Regression";24,1779059399203;4,08805293219918
+"PUPA";"Linear_Regression";22,6236119548177;3,6369652237688
+"CHMI";"Linear_Regression";24,0652256318907;3,27620518871196
+"SOAS";"Linear_Regression";21,3809744022612;55,9774194301704
+"PAIN";"Linear_Regression";26,5772382701118;2,12185549002321
+"SPRU";"Random_Forest";2,97499959615457;1,13107401451032
+"MEPO";"Random_Forest";7,48369664106253;0,392815379448158
+"FRPU";"Random_Forest";8,70361398694954;0,443703830474006
+"MEEL";"Random_Forest";9,06102286097607;0,516932050699917
+"POMA";"Random_Forest";12,4367911981773;0,311608832228401
+"HOMA";"Random_Forest";48,227831563224;0,120583607525557
+"SUSP";"Random_Forest";3,37906447992426;22,297797946556
+"RAPE";"Random_Forest";5,73065229482965;0,269184899412391
+"POMO";"Random_Forest";5,58847849780077;0,364809904145376
+"SCLA";"Random_Forest";3,12964922449755;0,580839601396784
+"SASO";"Random_Forest";14,2991816712285;0,294709082343761
+"CHFU";"Random_Forest";9,10515557349559;0,20663105301953
+"PLCO";"Random_Forest";12,6975234465724;0,357594475915358
+"BEMA";"Random_Forest";3,21728852623845;0,411685462782386
+"CETE";"Random_Forest";4,10503624545782;0,340506975214164
+"LEMA";"Random_Forest";24,9009121123238;0,223429359884701
+"MESU";"Random_Forest";8,02736862283518;0,447928142968908
+"LYTR";"Random_Forest";9,9674470125347;0,397880094909309
+"COSQ";"Random_Forest";7,03053529082599;0,34566541429772
+"PUPA";"Random_Forest";7,74114897540748;0,425820056269536
+"CHMI";"Random_Forest";7,5863193945307;0,32557681047917
+"SOAS";"Random_Forest";2,70141869687205;0,893596973244653
+"PAIN";"Random_Forest";10,3517676667916;0,321903082345207
+"SPRU";"XGBoost";3,86438002383194;1,90843274502541
+"MEPO";"XGBoost";8,0424279576535;0,453660003878899
+"FRPU";"XGBoost";9,66762679763936;0,547436422426238
+"MEEL";"XGBoost";9,16322609539304;0,528659222497529
+"POMA";"XGBoost";12,9022506473256;0,335369854026436
+"HOMA";"XGBoost";49,2040045883388;0,12551444276716
+"SUSP";"XGBoost";3,38445732409503;22,3690274280856
+"RAPE";"XGBoost";6,62117574299669;0,359345993314537
+"POMO";"XGBoost";6,90668696078156;0,557210121655775
+"SCLA";"XGBoost";4,53166641456801;1,21781342383957
+"SASO";"XGBoost";15,865672056352;0,36281739321417
+"CHFU";"XGBoost";11,0775004906507;0,305847125706849
+"PLCO";"XGBoost";13,5958438846592;0,409982285913314
+"BEMA";"XGBoost";5,96502268126685;1,41517357353932
+"CETE";"XGBoost";6,7730875872994;0,926969413287394
+"LEMA";"XGBoost";25,2958148118632;0,230572270785521
+"MESU";"XGBoost";8,87254849849105;0,547215890660385
+"LYTR";"XGBoost";10,3439669006473;0,428507656301279
+"COSQ";"XGBoost";8,29126455321431;0,480751544566285
+"PUPA";"XGBoost";8,95120407370413;0,569348497964712
+"CHMI";"XGBoost";8,98536553740819;0,456733342329019
+"SOAS";"XGBoost";3,74183304119022;1,71445739648376
+"PAIN";"XGBoost";11,4352016568458;0,392811106034513
